requirements.txt
numpy
torch==1.4.0
elementpath
optparse-pretty
python-resize-image
Pillow
torchvision
mahotas
matplotlib